+++ app/reflexes/portfolios_reflex.rb
@@ -3,19 +3,31 @@
 class PortfoliosReflex < ApplicationReflex
   def create(args={})
     create_portfolio(args['guid'])
-
-    @portfolios = Portfolios::Fetching::ForAccountService.call(user: current_user).result
-
     current_locale(args['locale'])
+    render_portfolios
   end
 
   private
+
+  def render_portfolios
+    morph(
+      '#portfolios',
+      PortfoliosController.render(
+        Portfolios::ListComponent.new(
+          current_user: current_user
+        )
+      )
+    )
+  end
 
   def create_portfolio(guid)
     Portfolios::CreateService.call(portfolio_params.merge(user: current_user, guid: guid))
   end
 
   def portfolio_params
-    params.require(:portfolio).permit(:name)
+    requets_params = params.require(:portfolio).permit(:name, :source, :currency).to_h.symbolize_keys
+    requets_params[:source] = requets_params[:source] == '-1' ? nil : requets_params[:source].to_i
+    requets_params[:currency] = requets_params[:currency].to_i
+    requets_params
   end
 end