+++ frontend/controllers/portfolios/cash_new_controller.js
@@ -4,6 +4,14 @@ export default class extends ApplicationController {
   submitForm (event) {
     event.preventDefault()
 
-    this.stimulate('Portfolios::CashesReflex#create', document.getElementById('current_locale').value)
+    this.stimulate(
+      'Portfolios::CashesReflex#create',
+      {
+        portfolio_id:   document.getElementById('portfolio_id').value,
+        show_plans:     document.getElementById('show_plans').value,
+        show_dividents: document.getElementById('show_dividents').value,
+        locale:         document.getElementById('current_locale').value
+      }
+    )
   }
 }